+++ .idea/deploymentTargetSelector.xml
@@ -4,10 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2025-02-19T09:46:32.340747667Z">
+        <DropdownSelection timestamp="2025-02-27T14:48:13.633244683Z">
           <Target type="DEFAULT_BOOT">
             <handle>
-              <DeviceId pluginId="PhysicalDevice" identifier="serial=R83W300AHTF" />
+              <DeviceId pluginId="LocalEmulator" identifier="path=/home/don/.android/avd/Medium_Phone.avd" />
             </handle>
           </Target>
         </DropdownSelection>
